Add files via upload
Here, Stakehoders and management want an insightful dashboard to know which customers needs to retain and the ones that needs to let go (the ones that are not profitable to the business) but the business is spending much on them. The dashboard is created with PowerBI and SQL is used.